Fix another clippy::unnecessary_cast
For some reason I don't see this warning on the bench.

+++ src/libxfuse/dir3_leaf.rs
@@ -155,7 +155,7 @@ impl<R: BufRead + Seek> Dir3<R> for Dir2Leaf {
         let mut entry: &Dir2Data = &self.entries[idx];
 
         buf_reader
-            .seek(SeekFrom::Start(entry.offset + (offset as u64)))
+            .seek(SeekFrom::Start(entry.offset + offset))
             .unwrap();
 
         loop {
